requirements.txt
discord.py==2.4.0
wavelink==3.4.1
